Give settled rows the full workspace context menu
Settled rows previously had no right-click menu at all — you had to
un-settle a row before you could archive, rename, or delete it. Both
inbox row shapes now share one menu wrapper (workspace-inbox-menu.tsx,
extracted from the card): the standard workspace actions plus a
Settle/Un-settle entry on top — 'Settle workspace' on cards (guardrail
permitting), 'Un-settle workspace' on settled rows. The card's inline
menu/dialog machinery moved into the shared wrapper unchanged.
